split new code into separate files a bit more
